Safari-609.1.11.1/LayoutTests/imported/w3c/web-platform-tests/webaudio/the-audio-api/the-convolvernode-interface/convolver-response-1-chan-expected.txt

CONSOLE MESSAGE: line 58: SyntaxError: The string did not match the expected pattern.

Harness Error (FAIL), message = SyntaxError: The string did not match the expected pattern.

PASS # AUDIT TASK RUNNER STARTED. 
PASS > [initialize] Convolver response with one channel 
FAIL X new AudioBuffer({numberOfChannels: 1, length: 2, sampleRate: 8192}) incorrectly threw TypeError: "function is not a constructor (evaluating 'new AudioBuffer(
                      {numberOfChannels: 1, length: 2, sampleRate: sampleRate})')". assert_true: expected true got false
FAIL < [initialize] 1 out of 1 assertions were failed. assert_true: expected true got false
PASS > [1-channel input] produces 1-channel output 
